--- a/user_tools/src/spark_rapids_pytools/common/sys_storage.py
+++ b/user_tools/src/spark_rapids_pytools/common/sys_storage.py
@@ -32,7 +32,7 @@ from fastcore.all import urlsave
 from fastprogress.fastprogress import progress_bar
 
 from spark_rapids_pytools.common.exceptions import StorageException
-from spark_rapids_pytools.common.utilities import Utils
+from spark_rapids_pytools.common.utilities import Utils, DownloaderVerification
 
 
 class FSUtil:
@@ -166,7 +166,10 @@ class FSUtil:
                 diff_time = int(datetime.datetime.now().timestamp() - modified_time)
                 if diff_time > expiration_time_s:
                     return False
-            # TODO verify using hashing
+            algorithm = DownloaderVerification.get_integrity_algorithm(checks_args)
+            # If valid algorithm present in check_args dictionary, verify integrity
+            if algorithm is not None:
+                return DownloaderVerification.check_integrity(fpath, algorithm, checks_args[algorithm])
             return True
 
         curr_time_stamp = datetime.datetime.now().timestamp()