Fixes #469.

### Change
Implemented a new module called `FileVerifier` to provide integrity checks for files using either GPG signature files or hashing techniques.

### Design
1. The module provides methods to get the signature file and/or hashing value.
2. It verifies the file's integrity using `gpg` for signature verification.
3. In case the `gpg` verification fails (gpg not installed, process timeout, import key error etc.), the module falls back to verification using hashing techniques.

### Performance Considerations
1. If the dependency file is not downloaded, the verification part will not be a bottleneck as majority of time will including downloading it.
2. GPG verification may be slower, particularly if the required public keys have not been imported.


### Timing Metrics
These metrics assume the jars have been downloaded before.  


#### Overall Time to Complete Dependencies Setup

| Environment   | With Verification (sec) | Without Verification (sec) |
|---------------|-------------------------|----------------------------|
| EMR           | 3.117                   | 1.838                      |
| Dataproc      | 2.870                   | 1.834                      |
| Databricks-AWS| 3.071                   | 1.827                      |

#### Verification Time for Individual Dependency

| Platform       | Dependency               | Verification Time (sec) | Size (MB) |
|----------------|--------------------------|------------------------|-------|
| EMR            | Hadoop AWS               | 0.005                  | 0.91 |
|                | AWS Java SDK Bundled     | 0.461                  | 268 |
|                | Apache Spark             | 0.482                  | 286 |
| Dataproc       | GCS Connector Hadoop3    | 0.085                  | 35 |
|                | Apache Spark             | 0.460                  | 286 |
| Databricks-AWS | Hadoop AWS               | 0.003                  | 0.91 |
|                | AWS Java SDK Bundled     | 0.417                  | 268 |
|                | Apache Spark             | 0.490                  | 286 |


#### System Configuration:
**CPU**: 36 cores Intel i9-10980XE @ 3.00GHz
**Memory**: 32 GB
**OS**: Ubuntu 22.04