@amahussein 2023-08-02
> Thanks @parthosa ! Overall design look good. Just wanted to know if you tested the performance impact while using GPG verification. Also the overall time taken while using GPG/hash vs without verifier.

I had problems running the wrapper with GPG verification.
This is probably something to do with how the gpg is installed and configured.
I find that since we cannot have the gpg-signature reliably working, we disable it and rely on hashlib as verification for now.

Hopefully @parthosa can give some quick numbers about the expected slow-down resulting from verifying the files.