@parthosa 2023-08-02
Below are the timing metrics for downloading dependencies with and without verification:

**Overall Time to Complete Downloading Dependencies**

| Environment   | With Verification (sec) | Without Verification (sec) |
|---------------|-------------------------|----------------------------|
| EMR           | 3.117                   | 1.838                      |
| Dataproc      | 2.870                   | 1.834                      |
| Databricks-aws| 3.071                   | 1.827                      |

**Verification Time for Individual Dependency**

| Platform       | Dependency               | Hashing - Cached (sec) | Size (MB) |
|----------------|--------------------------|------------------------|-------|
| EMR            | Hadoop AWS               | 0.005                  | 0.91 |
|                | AWS Java SDK Bundled     | 0.461                  | 268 |
|                | Apache Spark             | 0.482                  | 286 |
| Dataproc       | GCS Connector Hadoop3    | 0.085                  | 35 |
|                | Apache Spark             | 0.460                  | 286 |
| Databricks-AWS | Hadoop AWS               | 0.003                  | 0.91 |
|                | AWS Java SDK Bundled     | 0.417                  | 268 |
|                | Apache Spark             | 0.490                  | 286 |